refactor(wasm): extract IR dataclasses and runtime import tables
Continues the wasm.py → wasm/ package carve-up:

- _ir.py (~600 lines): ValType, SectionId, WasmOp enums; WasmInstruction,
  WasmFunction, WasmImport, WasmData, DiagSite, DiagMap, WasmModule
  dataclasses; _valtype_name / _format_wat_instr / _decode_leb128_* for
  WAT rendering. WAT formatting no longer has to reach into the emitter
  module.

- _imports.py (~440 lines): _RUNTIME_IMPORTS, _CMD_RUNTIME, the string
  and dict subcommand dispatch tables, _CLOCK_SUBCMD_IMPORT, the
  _SCOPE_NOP_COMMANDS and _UNSUPPORTED_COMMANDS sets, and the
  _OBJ_LIFECYCLE_IMPORTS always-imported set. Pure data, no runtime
  dependency on the emitter.

__init__.py drops from ~7720 to ~6590 lines and now re-imports these
symbols for internal use and for back-compat with external callers.

All 9985 prep-pr tests pass; lint and typecheck clean.
