+++ core/compiler/codegen/wasm/__init__.py
@@ -19,10 +19,8 @@ via :meth:`WasmModule.to_bytes` or inspected via :meth:`WasmModule.to_wat`.
 
 from __future__ import annotations
 
-import struct
 from collections.abc import Callable
 from dataclasses import dataclass, field
-from enum import IntEnum
 
 from ....analysis.semantic_model import Range
 from ....parsing.substitution import backslash_subst as _tcl_backslash_subst
@@ -93,1027 +91,38 @@ from ._encoding import (  # noqa: E402
     _tcl_token_value,
 )
 
-# WASM type constants
-
-
-class ValType(IntEnum):
-    """WASM value types."""
-
-    I32 = 0x7F
-    I64 = 0x7E
-    F32 = 0x7D
-    F64 = 0x7C
-
-
-class SectionId(IntEnum):
-    """WASM section identifiers."""
-
-    CUSTOM = 0
-    TYPE = 1
-    IMPORT = 2
-    FUNCTION = 3
-    TABLE = 4
-    MEMORY = 5
-    GLOBAL = 6
-    EXPORT = 7
-    START = 8
-    ELEMENT = 9
-    CODE = 10
-    DATA = 11
-
-
-# WASM opcodes used in emission
-class WasmOp(IntEnum):
-    """WASM instruction opcodes (subset used by this backend)."""
-
-    UNREACHABLE = 0x00
-    NOP = 0x01
-    BLOCK = 0x02
-    LOOP = 0x03
-    IF = 0x04
-    ELSE = 0x05
-    END = 0x0B
-    BR = 0x0C
-    BR_IF = 0x0D
-    BR_TABLE = 0x0E
-    RETURN = 0x0F
-    CALL = 0x10
-    DROP = 0x1A
-    SELECT = 0x1B
-    LOCAL_GET = 0x20
-    LOCAL_SET = 0x21
-    LOCAL_TEE = 0x22
-    GLOBAL_GET = 0x23
-    GLOBAL_SET = 0x24
-    I32_LOAD = 0x28
-    I64_LOAD = 0x29
-    I32_STORE = 0x36
-    I64_STORE = 0x37
-    MEMORY_SIZE = 0x3F
-    MEMORY_GROW = 0x40
-    I32_CONST = 0x41
-    I64_CONST = 0x42
-    F64_CONST = 0x44
-    I32_EQZ = 0x45
-    I32_EQ = 0x46
-    I32_NE = 0x47
-    I32_LT_S = 0x48
-    I32_GT_S = 0x4A
-    I32_LE_S = 0x4C
-    I32_GE_S = 0x4E
-    I64_EQZ = 0x50
-    I64_EQ = 0x51
-    I64_NE = 0x52
-    I64_LT_S = 0x53
-    I64_GT_S = 0x55
-    I64_LE_S = 0x57
-    I64_GE_S = 0x59
-    I32_ADD = 0x6A
-    I32_SUB = 0x6B
-    I32_MUL = 0x6C
-    I32_DIV_S = 0x6D
-    I32_REM_S = 0x6F
-    I32_AND = 0x71
-    I32_OR = 0x72
-    I32_XOR = 0x73
-    I32_SHL = 0x74
-    I32_SHR_S = 0x75
-    I64_ADD = 0x7C
-    I64_SUB = 0x7D
-    I64_MUL = 0x7E
-    I64_DIV_S = 0x7F
-    I64_REM_S = 0x81
-    I64_AND = 0x83
-    I64_OR = 0x84
-    I64_XOR = 0x85
-    I64_SHL = 0x86
-    I64_SHR_S = 0x87
-    I32_WRAP_I64 = 0xA7
-    I64_EXTEND_I32_S = 0xAC
-
-
-# WASM magic and version
-_WASM_MAGIC = b"\x00asm"
-_WASM_VERSION = struct.pack("<I", 1)
-
-# Block type for void blocks
-_BLOCK_VOID = 0x40
-# Block type for i64 result
-_BLOCK_I64 = ValType.I64
-
-
-# Data structures
-
-
-@dataclass(slots=True)
-class WasmInstruction:
-    """A single WASM instruction with optional operands."""
-
-    op: int
-    operands: bytes = b""
-
-    def encode(self) -> bytes:
-        return bytes([self.op]) + self.operands
-
-
-@dataclass(slots=True)
-class WasmFunction:
-    """Compiled WASM function."""
-
-    name: str
-    params: list[ValType]
-    results: list[ValType]
-    locals: list[ValType]
-    body: list[WasmInstruction]
-    local_names: list[str] = field(default_factory=list)
-    exported: bool = False
-
-    def encode_body(self) -> bytes:
-        """Encode function body (locals + instructions + end)."""
-        # Encode local declarations (run-length compressed)
-        local_groups: list[bytes] = []
-        if self.locals:
-            # Group consecutive same-type locals
-            groups: list[tuple[int, ValType]] = []
-            current_type = self.locals[0]
-            count = 1
-            for lt in self.locals[1:]:
-                if lt == current_type:
-                    count += 1
-                else:
-                    groups.append((count, current_type))
-                    current_type = lt
-                    count = 1
-            groups.append((count, current_type))
-            for cnt, vt in groups:
-                local_groups.append(_leb128_unsigned(cnt) + bytes([vt]))
-
-        locals_section = _encode_vector(local_groups)
-        code = b"".join(instr.encode() for instr in self.body)
-        # Ensure the body ends with END
-        if not self.body or self.body[-1].op != WasmOp.END:
-            code += bytes([WasmOp.END])
-        func_body = locals_section + code
-        return _leb128_unsigned(len(func_body)) + func_body
-
-
-@dataclass(slots=True)
-class WasmImport:
-    """An imported function."""
-
-    module: str
-    name: str
-    type_idx: int
-
-
-@dataclass(slots=True)
-class WasmData:
-    """A data segment for string constants."""
-
-    offset: int
-    data: bytes
-
-
-@dataclass(slots=True)
-class DiagSite:
-    """A single diagnostic call-site in the compiled module.
-
-    Emitted by the codegen as a sidecar entry for every command invocation
-    whose trap path surfaces the command to the user — the eval fallback,
-    unsupported-command trap, and unknown-command dispatch.  At runtime
-    the last-registered site's ID is prefixed onto the stderr trap line
-    so a companion resolver (tcl-trap-resolve skill, or an in-process
-    harness loader) can map ``site=1234`` back to ``(file, line, col,
-    command)`` without embedding the strings in the .wasm.
-    """
-
-    id: int  # site_id emitted as the argument to tcl_diag_set
-    file: str  # source filename (may be "<unknown>" when not threaded)
-    line: int  # 1-based line number of the originating command
-    col: int  # 1-based column
-    end_line: int
-    end_col: int
-    command: str  # the Tcl command being dispatched
-    args: tuple[str, ...] = ()  # raw argument strings (best-effort)
-    kind: str = "fallback"  # "fallback" | "unsupported" | "unknown"
-    proc: str | None = None  # fully-qualified proc name this site lives in
-
-
-@dataclass(slots=True)
-class DiagMap:
-    """Sidecar source-location map for a compiled WASM module.
-
-    Produced alongside a :class:`WasmModule` by
-    :func:`wasm_codegen_module` when a filename is supplied.  The map
-    resolves the opaque ``site_id`` values that the runtime writes on a
-    trap back to concrete source locations.  :func:`to_json_dict` emits
-    the wire format that the runner writes to ``<module>.wasm.map.json``.
-
-    ``procs`` separately resolves WASM function indices to their proc
-    qualified names so a wasm backtrace like ``<wasm function 73>`` can
-    be rendered as ``::tcltest::Configure``.
-    """
-
-    sites: list[DiagSite] = field(default_factory=list)
-    procs: list[tuple[int, str]] = field(default_factory=list)
-    filename: str = "<unknown>"
-
-    def add_site(self, site: DiagSite) -> int:
-        """Register a site and return its id (1-based; 0 means 'unset')."""
-        self.sites.append(site)
-        return site.id
-
-    def to_json_dict(self) -> dict:
-        return {
-            "filename": self.filename,
-            "sites": [
-                {
-                    "id": s.id,
-                    "file": s.file,
-                    "line": s.line,
-                    "col": s.col,
-                    "end_line": s.end_line,
-                    "end_col": s.end_col,
-                    "command": s.command,
-                    "args": list(s.args),
-                    "kind": s.kind,
-                    "proc": s.proc,
-                }
-                for s in self.sites
-            ],
-            "procs": [{"func_idx": f, "qname": q} for f, q in self.procs],
-        }
-
-
-@dataclass(slots=True)
-class WasmModule:
-    """Complete WASM module."""
-
-    functions: list[WasmFunction] = field(default_factory=list)
-    imports: list[WasmImport] = field(default_factory=list)
-    data_segments: list[WasmData] = field(default_factory=list)
-    memory_pages: int = 1
-    import_memory: bool = True  # import memory from runtime instead of defining own
-    _type_cache: dict[tuple[tuple[int, ...], tuple[int, ...]], int] = field(default_factory=dict)
-    _types: list[tuple[list[ValType], list[ValType]]] = field(default_factory=list)
-
-    def _intern_type(self, params: list[ValType], results: list[ValType]) -> int:
-        key = (tuple(params), tuple(results))
-        idx = self._type_cache.get(key)
-        if idx is not None:
-            return idx
-        idx = len(self._types)
-        self._types.append((list(params), list(results)))
-        self._type_cache[key] = idx
-        return idx
-
-    def to_bytes(self) -> bytes:
-        """Serialise to a valid WASM binary."""
-        sections: list[bytes] = []
-
-        # Register all function types
-        for imp in self.imports:
-            # Import types already registered
-            pass
-        for func in self.functions:
-            self._intern_type(func.params, func.results)
-
-        # Type section
-        type_entries: list[bytes] = []
-        for params, results in self._types:
-            entry = b"\x60"  # functype
-            entry += _encode_vector([bytes([p]) for p in params])
-            entry += _encode_vector([bytes([r]) for r in results])
-            type_entries.append(entry)
-        if type_entries:
-            sections.append(self._make_section(SectionId.TYPE, _encode_vector(type_entries)))
-
-        # Import section
-        if self.imports:
-            import_entries: list[bytes] = []
-            for imp in self.imports:
-                entry = _encode_string(imp.module)
-                entry += _encode_string(imp.name)
-                entry += b"\x00"  # functype
-                entry += _leb128_unsigned(imp.type_idx)
-                import_entries.append(entry)
-            # Import memory from runtime so data segments share the same
-            # linear memory as TclObj allocations.
-            if self.import_memory:
-                mem_imp = _encode_string("tcl")
-                mem_imp += _encode_string("memory")
-                mem_imp += b"\x02"  # memory import
-                mem_imp += b"\x00"  # limits: no maximum
-                mem_imp += _leb128_unsigned(self.memory_pages)
-                import_entries.append(mem_imp)
-            sections.append(self._make_section(SectionId.IMPORT, _encode_vector(import_entries)))
-
-        # Function section (type indices for defined functions)
-        if self.functions:
-            func_type_indices: list[bytes] = []
-            for func in self.functions:
-                tidx = self._intern_type(func.params, func.results)
-                func_type_indices.append(_leb128_unsigned(tidx))
-            sections.append(
-                self._make_section(SectionId.FUNCTION, _encode_vector(func_type_indices))
-            )
-
-        # Memory section (only if not importing)
-        if not self.import_memory:
-            mem_data = _leb128_unsigned(1)  # 1 memory
-            mem_data += b"\x00"  # no maximum
-            mem_data += _leb128_unsigned(self.memory_pages)
-            sections.append(self._make_section(SectionId.MEMORY, mem_data))
-
-        # Export section
-        export_entries: list[bytes] = []
-        # Export memory (only if we define it, not if imported)
-        if not self.import_memory:
-            exp = _encode_string("memory")
-            exp += b"\x02"  # memory export
-            exp += _leb128_unsigned(0)
-            export_entries.append(exp)
-        # Export functions
-        num_imports = len(self.imports)
-        for i, func in enumerate(self.functions):
-            if func.exported:
-                exp = _encode_string(func.name)
-                exp += b"\x00"  # func export
-                exp += _leb128_unsigned(num_imports + i)
-                export_entries.append(exp)
-        if export_entries:
-            sections.append(self._make_section(SectionId.EXPORT, _encode_vector(export_entries)))
-
-        # Code section
-        if self.functions:
-            code_entries = [func.encode_body() for func in self.functions]
-            sections.append(self._make_section(SectionId.CODE, _encode_vector(code_entries)))
-
-        # Data section
-        if self.data_segments:
-            data_entries: list[bytes] = []
-            for seg in self.data_segments:
-                entry = b"\x00"  # active, memory 0
-                # i32.const offset + end
-                entry += bytes([WasmOp.I32_CONST]) + _leb128_signed(seg.offset)
-                entry += bytes([WasmOp.END])
-                entry += _leb128_unsigned(len(seg.data)) + seg.data
-                data_entries.append(entry)
-            sections.append(self._make_section(SectionId.DATA, _encode_vector(data_entries)))
-
-        return _WASM_MAGIC + _WASM_VERSION + b"".join(sections)
-
-    def to_wat(self) -> str:
-        """Generate a human-readable WAT (WebAssembly Text) representation."""
-        lines: list[str] = ["(module"]
-
-        # Types
-        for i, (params, results) in enumerate(self._types):
-            p = " ".join(f"(param {_valtype_name(p)})" for p in params)
-            r = " ".join(f"(result {_valtype_name(r)})" for r in results)
-            lines.append(f"  (type $t{i} (func {p} {r}))")
-
-        # Imports
-        for imp in self.imports:
-            lines.append(
-                f'  (import "{imp.module}" "{imp.name}" '
-                f"(func $imp_{imp.name} (type $t{imp.type_idx})))"
-            )
-
-        # Memory
-        if self.import_memory:
-            lines.append(f'  (import "tcl" "memory" (memory {self.memory_pages}))')
-        else:
-            lines.append(f'  (memory (export "memory") {self.memory_pages})')
-
-        # Functions
-        for func in self.functions:
-            self._intern_type(func.params, func.results)
-            export = f' (export "{func.name}")' if func.exported else ""
-            sig_parts: list[str] = []
-            for j, p in enumerate(func.params):
-                name = func.local_names[j] if j < len(func.local_names) else f"$p{j}"
-                sig_parts.append(f"(param {name} {_valtype_name(p)})")
-            for r in func.results:
-                sig_parts.append(f"(result {_valtype_name(r)})")
-            sig = " ".join(sig_parts)
-
-            lines.append(f"  (func ${func.name}{export} {sig}")
-
-            # Locals
-            param_count = len(func.params)
-            for j, lt in enumerate(func.locals):
-                local_idx = param_count + j
-                name = (
-                    func.local_names[local_idx] if local_idx < len(func.local_names) else f"$l{j}"
-                )
-                lines.append(f"    (local {name} {_valtype_name(lt)})")
-
-            # Body
-            indent = 2
-            for instr in func.body:
-                line = _format_wat_instr(instr, indent)
-                if line is not None:
-                    lines.append(line)
-                    # Track indent for blocks
-                    if instr.op in (WasmOp.BLOCK, WasmOp.LOOP, WasmOp.IF):
-                        indent += 1
-                    elif instr.op == WasmOp.ELSE:
-                        pass  # same indent
-                    elif instr.op == WasmOp.END:
-                        indent = max(2, indent - 1)
-
-            lines.append("  )")
-
-        # Data segments
-        for seg in self.data_segments:
-            escaped = (
-                seg.data.decode("utf-8", errors="replace").replace("\\", "\\\\").replace('"', '\\"')
-            )
-            lines.append(f'  (data (i32.const {seg.offset}) "{escaped}")')
-
-        lines.append(")")
-        return "\n".join(lines)
-
-    @staticmethod
-    def _make_section(section_id: SectionId, content: bytes) -> bytes:
-        return bytes([section_id]) + _leb128_unsigned(len(content)) + content
-
-
-def _valtype_name(vt: ValType | int) -> str:
-    names = {ValType.I32: "i32", ValType.I64: "i64", ValType.F32: "f32", ValType.F64: "f64"}
-    return names.get(ValType(vt), f"unknown({vt})")
-
-
-def _format_wat_instr(instr: WasmInstruction, indent: int) -> str | None:
-    """Format a single instruction for WAT output."""
-    prefix = "    " * indent
-    op = instr.op
-
-    _WAT_NAMES: dict[int, str] = {
-        WasmOp.UNREACHABLE: "unreachable",
-        WasmOp.NOP: "nop",
-        WasmOp.BLOCK: "block",
-        WasmOp.LOOP: "loop",
-        WasmOp.IF: "if",
-        WasmOp.ELSE: "else",
-        WasmOp.END: "end",
-        WasmOp.BR: "br",
-        WasmOp.BR_IF: "br_if",
-        WasmOp.BR_TABLE: "br_table",
-        WasmOp.RETURN: "return",
-        WasmOp.CALL: "call",
-        WasmOp.DROP: "drop",
-        WasmOp.SELECT: "select",
-        WasmOp.LOCAL_GET: "local.get",
-        WasmOp.LOCAL_SET: "local.set",
-        WasmOp.LOCAL_TEE: "local.tee",
-        WasmOp.GLOBAL_GET: "global.get",
-        WasmOp.GLOBAL_SET: "global.set",
-        WasmOp.I32_CONST: "i32.const",
-        WasmOp.I64_CONST: "i64.const",
-        WasmOp.F64_CONST: "f64.const",
-        WasmOp.I32_EQZ: "i32.eqz",
-        WasmOp.I32_EQ: "i32.eq",
-        WasmOp.I32_NE: "i32.ne",
-        WasmOp.I32_LT_S: "i32.lt_s",
-        WasmOp.I32_GT_S: "i32.gt_s",
-        WasmOp.I32_LE_S: "i32.le_s",
-        WasmOp.I32_GE_S: "i32.ge_s",
-        WasmOp.I64_EQZ: "i64.eqz",
-        WasmOp.I64_EQ: "i64.eq",
-        WasmOp.I64_NE: "i64.ne",
-        WasmOp.I64_LT_S: "i64.lt_s",
-        WasmOp.I64_GT_S: "i64.gt_s",
-        WasmOp.I64_LE_S: "i64.le_s",
-        WasmOp.I64_GE_S: "i64.ge_s",
-        WasmOp.I32_ADD: "i32.add",
-        WasmOp.I32_SUB: "i32.sub",
-        WasmOp.I32_MUL: "i32.mul",
-        WasmOp.I32_DIV_S: "i32.div_s",
-        WasmOp.I32_REM_S: "i32.rem_s",
-        WasmOp.I32_AND: "i32.and",
-        WasmOp.I32_OR: "i32.or",
-        WasmOp.I32_XOR: "i32.xor",
-        WasmOp.I32_SHL: "i32.shl",
-        WasmOp.I32_SHR_S: "i32.shr_s",
-        WasmOp.I64_ADD: "i64.add",
-        WasmOp.I64_SUB: "i64.sub",
-        WasmOp.I64_MUL: "i64.mul",
-        WasmOp.I64_DIV_S: "i64.div_s",
-        WasmOp.I64_REM_S: "i64.rem_s",
-        WasmOp.I64_AND: "i64.and",
-        WasmOp.I64_OR: "i64.or",
-        WasmOp.I64_XOR: "i64.xor",
-        WasmOp.I64_SHL: "i64.shl",
-        WasmOp.I64_SHR_S: "i64.shr_s",
-        WasmOp.I32_WRAP_I64: "i32.wrap_i64",
-        WasmOp.I64_EXTEND_I32_S: "i64.extend_i32_s",
-        WasmOp.I32_STORE: "i32.store",
-        WasmOp.I64_STORE: "i64.store",
-        WasmOp.I32_LOAD: "i32.load",
-        WasmOp.I64_LOAD: "i64.load",
-        WasmOp.MEMORY_SIZE: "memory.size",
-        WasmOp.MEMORY_GROW: "memory.grow",
-    }
-
-    name = _WAT_NAMES.get(op)
-    if name is None:
-        return f"{prefix};; unknown op 0x{op:02x}"
-
-    # Decode operands for display
-    if op in (
-        WasmOp.LOCAL_GET,
-        WasmOp.LOCAL_SET,
-        WasmOp.LOCAL_TEE,
-        WasmOp.GLOBAL_GET,
-        WasmOp.GLOBAL_SET,
-        WasmOp.CALL,
-        WasmOp.BR,
-        WasmOp.BR_IF,
-    ):
-        if instr.operands:
-            # Decode LEB128 index from operands
-            idx = _decode_leb128_unsigned(instr.operands)
-            return f"{prefix}{name} {idx}"
-    elif op in (WasmOp.I32_CONST,):
-        if instr.operands:
-            val = _decode_leb128_signed(instr.operands)
-            return f"{prefix}{name} {val}"
-    elif op in (WasmOp.I64_CONST,):
-        if instr.operands:
-            val = _decode_leb128_signed(instr.operands)
-            return f"{prefix}{name} {val}"
-    elif op == WasmOp.BLOCK or op == WasmOp.LOOP or op == WasmOp.IF:
-        if instr.operands:
-            bt = instr.operands[0]
-            if bt == _BLOCK_VOID:
-                return f"{prefix}{name}"
-            return f"{prefix}{name} (result {_valtype_name(bt)})"
-        return f"{prefix}{name}"
-
-    return f"{prefix}{name}"
-
-
-def _decode_leb128_unsigned(data: bytes) -> int:
-    result = 0
-    shift = 0
-    for byte in data:
-        result |= (byte & 0x7F) << shift
-        if not (byte & 0x80):
-            break
-        shift += 7
-    return result
-
-
-def _decode_leb128_signed(data: bytes) -> int:
-    result = 0
-    shift = 0
-    for byte in data:
-        result |= (byte & 0x7F) << shift
-        shift += 7
-        if not (byte & 0x80):
-            if byte & 0x40:
-                result -= 1 << shift
-            break
-    return result
-
-
 # WASM Emitter
-
-
-# Runtime function signatures imported from the Tcl runtime.
-# Each entry maps an import key to (module, export_name, param_types, result_types).
-#
-# Values are represented as i32 TclObj pointers.  The TclObj lifecycle
-# imports (obj_new_int, obj_new_string, obj_get_int) bridge between raw
-# WASM integers (i64) used in expr arithmetic and the i32 object pointers
-# passed to/from runtime functions and stored in locals.
-_RUNTIME_IMPORTS: dict[str, tuple[str, str, list[ValType], list[ValType]]] = {
-    # TclObj lifecycle — always imported
-    "tcl_obj_new_int": ("tcl", "obj_new_int", [ValType.I64], [ValType.I32]),
-    "tcl_obj_new_string": ("tcl", "obj_new_string", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_obj_get_int": ("tcl", "obj_get_int", [ValType.I32], [ValType.I64]),
-    # Command runtime — all parameters/results are i32 TclObj pointers
-    "tcl_puts": ("tcl", "tcl_cmd_puts", [ValType.I32], [ValType.I32]),
-    "tcl_append": ("tcl", "tcl_cmd_append", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_list_length": ("tcl", "tcl_cmd_list_length", [ValType.I32], [ValType.I32]),
-    "tcl_lappend": ("tcl", "tcl_cmd_lappend", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_length": ("tcl", "string_length", [ValType.I32], [ValType.I32]),
-    "tcl_string_index": ("tcl", "string_index", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_range": (
-        "tcl",
-        "string_range",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_string_compare": (
-        "tcl",
-        "string_compare",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_expr_order_cmp": (
-        "tcl",
-        "tcl_expr_order_cmp",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_string_map": ("tcl", "string_map", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_match": ("tcl", "string_match", [ValType.I32, ValType.I32], [ValType.I32]),
-    # ``string trim{,left,right} value ?chars?`` — a null (i32 0) chars arg
-    # means "default whitespace"; otherwise the TclObj's string value is the
-    # set of bytes to trim.
-    "tcl_string_trim": ("tcl", "string_trim", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_trimleft": (
-        "tcl",
-        "string_trimleft",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_string_trimright": (
-        "tcl",
-        "string_trimright",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_string_equal": ("tcl", "string_equal", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_first": ("tcl", "string_first", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_last": ("tcl", "string_last", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_repeat": ("tcl", "string_repeat", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_string_reverse": ("tcl", "string_reverse", [ValType.I32], [ValType.I32]),
-    "tcl_string_toupper": ("tcl", "string_toupper", [ValType.I32], [ValType.I32]),
-    "tcl_string_tolower": ("tcl", "string_tolower", [ValType.I32], [ValType.I32]),
-    "tcl_string_replace": (
-        "tcl",
-        "string_replace",
-        [ValType.I32, ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_string_is_integer": ("tcl", "string_is_integer", [ValType.I32], [ValType.I32]),
-    "tcl_string_is_alpha": ("tcl", "string_is_alpha", [ValType.I32], [ValType.I32]),
-    "tcl_string_is_digit": ("tcl", "string_is_digit", [ValType.I32], [ValType.I32]),
-    "tcl_string_is_space": ("tcl", "string_is_space", [ValType.I32], [ValType.I32]),
-    "tcl_list_create": ("tcl", "tcl_list", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_concat": ("tcl", "tcl_cmd_concat", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_list_index": ("tcl", "tcl_cmd_list_index", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_list_range": (
-        "tcl",
-        "tcl_cmd_list_range",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_list_tail": ("tcl", "list_tail", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_list_sort": ("tcl", "tcl_cmd_list_sort", [ValType.I32], [ValType.I32]),
-    "tcl_list_search": ("tcl", "tcl_cmd_list_search", [ValType.I32, ValType.I32], [ValType.I32]),
-    # Dict commands
-    "tcl_dict_create": ("tcl", "dict_create", [], [ValType.I32]),
-    "tcl_dict_get": ("tcl", "dict_get", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_dict_set": (
-        "tcl",
-        "dict_set",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_dict_exists": ("tcl", "dict_exists", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_dict_keys": ("tcl", "dict_keys", [ValType.I32], [ValType.I32]),
-    "tcl_dict_values": ("tcl", "dict_values", [ValType.I32], [ValType.I32]),
-    "tcl_dict_size": ("tcl", "dict_size", [ValType.I32], [ValType.I32]),
-    "tcl_error": ("tcl", "tcl_cmd_error", [ValType.I32], []),
-    "tcl_format": (
-        "tcl",
-        "tcl_cmd_format",
-        [ValType.I32, ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_regexp": ("tcl", "tcl_cmd_regexp", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_open": ("tcl", "tcl_cmd_open", [ValType.I32], [ValType.I32]),
-    "tcl_close": ("tcl", "tcl_cmd_close", [ValType.I32], [ValType.I32]),
-    "tcl_read": ("tcl", "tcl_cmd_read", [ValType.I32], [ValType.I32]),
-    "tcl_gets": ("tcl", "tcl_cmd_gets", [ValType.I32], [ValType.I32]),
-    # Global variable table
-    "tcl_global_set": ("tcl", "global_set", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_global_get": ("tcl", "global_get", [ValType.I32], [ValType.I32]),
-    "tcl_global_exists": ("tcl", "global_exists", [ValType.I32], [ValType.I32]),
-    # Catch / error handling
-    "tcl_catch_enter": ("tcl", "catch_enter", [], []),
-    "tcl_catch_leave": ("tcl", "catch_leave", [], [ValType.I32]),
-    "tcl_catch_result": ("tcl", "catch_result", [], [ValType.I32]),
-    "tcl_catch_has_error": ("tcl", "catch_has_error", [], [ValType.I32]),
-    "tcl_catch_set_ok_result": ("tcl", "catch_set_ok_result", [ValType.I32], []),
-    # Interpreter fallback
-    "tcl_eval": ("tcl", "tcl_eval", [ValType.I32], [ValType.I32]),
-    # Namespace context for eval-fallback calls — compiled procs
-    # set the current namespace before ``tcl_eval`` so dynamic
-    # ``proc $name`` / ``variable $name`` inside the fallback
-    # qualify into the enclosing namespace instead of the global
-    # scope.  ``ns_set`` returns an opaque i64 save token;
-    # ``ns_restore`` unwinds it.
-    "tcl_ns_set": (
-        "tcl",
-        "ns_set",
-        [ValType.I32, ValType.I32],
-        [ValType.I64],
-    ),
-    "tcl_ns_restore": ("tcl", "ns_restore", [ValType.I64], []),
-    # Frame stack (local variable scoping)
-    "tcl_frame_push": ("tcl", "frame_push", [], [ValType.I32]),
-    "tcl_frame_pop": ("tcl", "frame_pop", [], []),
-    "tcl_var_resolve": ("tcl", "var_resolve", [ValType.I32], [ValType.I32]),
-    "tcl_var_set": ("tcl", "var_set", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_var_exists": ("tcl", "var_exists", [ValType.I32], [ValType.I32]),
-    "tcl_local_set": ("tcl", "local_set", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_local_get": ("tcl", "local_get", [ValType.I32], [ValType.I32]),
-    # Proc registry
-    "tcl_proc_register": (
-        "tcl",
-        "proc_register",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_proc_register_compiled": (
-        "tcl",
-        "proc_register_compiled",
-        [ValType.I32, ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    # Info command
-    "tcl_info_exists": ("tcl", "info_exists", [ValType.I32], [ValType.I32]),
-    "tcl_info_dispatch": ("tcl", "info_dispatch", [ValType.I32, ValType.I32], [ValType.I32]),
-    # Clock command — WASI clock_time_get wrappers, integer results.
-    "tcl_clock_seconds": ("tcl", "clock_seconds", [], [ValType.I32]),
-    "tcl_clock_clicks": ("tcl", "clock_clicks", [], [ValType.I32]),
-    "tcl_clock_milliseconds": ("tcl", "clock_milliseconds", [], [ValType.I32]),
-    # Frame-depth helpers for uplevel — temporarily shift frame_depth
-    # so a called ``tcl_eval`` runs at a caller's scope.
-    "tcl_frame_depth_stash": (
-        "tcl",
-        "frame_depth_stash",
-        [ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_frame_depth_restore": ("tcl", "frame_depth_restore", [ValType.I32], []),
-    # Arrays — dedicated per-array hash tables.
-    "tcl_array_set": (
-        "tcl",
-        "array_set",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_array_get": ("tcl", "array_get", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_array_exists": ("tcl", "array_exists", [ValType.I32], [ValType.I32]),
-    "tcl_array_element_exists": (
-        "tcl",
-        "array_element_exists",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_array_size": ("tcl", "array_size", [ValType.I32], [ValType.I32]),
-    "tcl_array_unset": ("tcl", "array_unset", [ValType.I32], [ValType.I32]),
-    "tcl_array_unset_element": (
-        "tcl",
-        "array_unset_element",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_array_names": (
-        "tcl",
-        "array_names",
-        [ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    # String split/join
-    "tcl_split": ("tcl", "tcl_cmd_split", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_join": ("tcl", "tcl_cmd_join", [ValType.I32, ValType.I32], [ValType.I32]),
-    # Diagnostic — record the current source site so trap paths can
-    # prefix stderr with ``site=<id>`` for sidecar-map resolution.
-    "tcl_diag_set": ("tcl", "diag_set", [ValType.I32], []),
-    # I/O + channel stubs (tcl_io_stubs.zig) — trap with "unsupported
-    # command: <name>".  Each stub takes i32 args and returns i32.
-    "tcl_eof": ("tcl", "tcl_cmd_eof", [ValType.I32], [ValType.I32]),
-    "tcl_flush": ("tcl", "tcl_cmd_flush", [ValType.I32], [ValType.I32]),
-    "tcl_fblocked": ("tcl", "tcl_cmd_fblocked", [ValType.I32], [ValType.I32]),
-    "tcl_fconfigure": ("tcl", "tcl_cmd_fconfigure", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_tell": ("tcl", "tcl_cmd_tell", [ValType.I32], [ValType.I32]),
-    "tcl_seek": ("tcl", "tcl_cmd_seek", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_chan": ("tcl", "tcl_cmd_chan", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_fcopy": ("tcl", "tcl_cmd_fcopy", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_fileevent": ("tcl", "tcl_cmd_fileevent", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_socket": ("tcl", "tcl_cmd_socket", [ValType.I32, ValType.I32], [ValType.I32]),
-    # Filesystem / process stubs (tcl_fs_stubs.zig).
-    "tcl_file": (
-        "tcl",
-        "tcl_cmd_file",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    "tcl_glob": ("tcl", "tcl_cmd_glob", [ValType.I32], [ValType.I32]),
-    "tcl_pwd": ("tcl", "tcl_cmd_pwd", [], [ValType.I32]),
-    "tcl_cd": ("tcl", "tcl_cmd_cd", [ValType.I32], [ValType.I32]),
-    "tcl_exec": ("tcl", "tcl_cmd_exec", [ValType.I32], [ValType.I32]),
-    "tcl_source": ("tcl", "tcl_cmd_source", [ValType.I32], [ValType.I32]),
-    "tcl_load": ("tcl", "tcl_cmd_load", [ValType.I32], [ValType.I32]),
-    "tcl_unload": ("tcl", "tcl_cmd_unload", [ValType.I32], [ValType.I32]),
-    # Format / regex / encoding stubs (tcl_fmt_stubs.zig).
-    "tcl_scan": ("tcl", "tcl_cmd_scan", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_binary": ("tcl", "tcl_cmd_binary", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_regsub": ("tcl", "tcl_cmd_regsub", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_encoding": (
-        "tcl",
-        "tcl_cmd_encoding",
-        [ValType.I32, ValType.I32, ValType.I32],
-        [ValType.I32],
-    ),
-    # Time / event stubs (tcl_time_stubs.zig).  clock_format /
-    # clock_scan / clock_add are the non-implemented clock
-    # subcommands; the arithmetic ones (seconds / clicks /
-    # milliseconds) are real runtime fns in tcl_clock.zig.
-    "tcl_clock_format": ("tcl", "clock_format", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_clock_scan": ("tcl", "clock_scan", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_clock_add": ("tcl", "clock_add", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_after": ("tcl", "tcl_cmd_after", [ValType.I32], [ValType.I32]),
-    "tcl_vwait": ("tcl", "tcl_cmd_vwait", [ValType.I32], [ValType.I32]),
-    "tcl_update": ("tcl", "tcl_cmd_update", [], [ValType.I32]),
-    "tcl_coroutine": ("tcl", "tcl_cmd_coroutine", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_yield": ("tcl", "tcl_cmd_yield", [ValType.I32], [ValType.I32]),
-    "tcl_yieldto": ("tcl", "tcl_cmd_yieldto", [ValType.I32], [ValType.I32]),
-    # Environment / metadata stubs (tcl_env_stubs.zig).  ``namespace``
-    # and ``namespace eval`` are handled separately by the compiler;
-    # this stub catches the other namespace subcommands falling
-    # through.
-    "tcl_namespace": ("tcl", "namespace", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_package": ("tcl", "tcl_cmd_package_cmd", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_trace": ("tcl", "tcl_cmd_trace_cmd", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_interp": ("tcl", "tcl_cmd_interp_cmd", [ValType.I32, ValType.I32], [ValType.I32]),
-    "tcl_apply": ("tcl", "tcl_cmd_apply", [ValType.I32, ValType.I32], [ValType.I32]),
-}
-
-# Import keys for the TclObj lifecycle functions — always registered
-_OBJ_LIFECYCLE_IMPORTS = frozenset(
-    {
-        "tcl_obj_new_int",
-        "tcl_obj_new_string",
-        "tcl_obj_get_int",
-    }
+from ._imports import (  # noqa: E402
+    _CLOCK_SUBCMD_IMPORT,
+    _CMD_RUNTIME,
+    _DICT_SUBCMD_IMPORT,
+    _OBJ_LIFECYCLE_IMPORTS,
+    _RUNTIME_IMPORTS,
+    _SCOPE_NOP_COMMANDS,
+    _STRING_IS_IMPORT,
+    _STRING_SUBCMD_IMPORT,
+    _UNSUPPORTED_COMMANDS,
 )
-
-# Commands that map to runtime imports.  The value is
-# (import_key, arg_count_or_None) where ``None`` means variadic.
-#
-# Stub-mapped commands (file, glob, encoding, trace, etc.) route
-# through this dispatch table so the compiler calls the stub
-# directly — producing a clean ``unsupported command: <name>`` trap
-# — instead of generating a ``tcl_eval`` fallback that invokes the
-# interpreter which then fails to find the command.
-_CMD_RUNTIME: dict[str, tuple[str, int | None]] = {
-    # Implemented runtime functions
-    "puts": ("tcl_puts", 1),
-    "append": ("tcl_append", 2),
-    "llength": ("tcl_list_length", 1),
-    "lappend": ("tcl_lappend", 2),
-    "lindex": ("tcl_list_index", 2),
-    "lrange": ("tcl_list_range", 3),
-    "lsort": ("tcl_list_sort", 1),
-    "lsearch": ("tcl_list_search", 2),
-    "concat": ("tcl_concat", 2),
-    "error": ("tcl_error", 1),
-    "split": ("tcl_split", 2),
-    "join": ("tcl_join", 2),
-    # I/O stubs — open/close/read/gets/eof/flush/fblocked/fconfigure/
-    # tell/seek/chan/fcopy/fileevent/socket all trap with
-    # "unsupported command: <name>".
-    "open": ("tcl_open", 1),
-    "close": ("tcl_close", 1),
-    "read": ("tcl_read", 1),
-    "gets": ("tcl_gets", 1),
-    "eof": ("tcl_eof", 1),
-    "flush": ("tcl_flush", 1),
-    "fblocked": ("tcl_fblocked", 1),
-    "fconfigure": ("tcl_fconfigure", 2),
-    "tell": ("tcl_tell", 1),
-    "seek": ("tcl_seek", 2),
-    "chan": ("tcl_chan", 2),
-    "fcopy": ("tcl_fcopy", 2),
-    "fileevent": ("tcl_fileevent", 2),
-    "socket": ("tcl_socket", 2),
-    # Filesystem / process stubs.
-    "file": ("tcl_file", 3),
-    "glob": ("tcl_glob", 1),
-    "pwd": ("tcl_pwd", 0),
-    "cd": ("tcl_cd", 1),
-    "exec": ("tcl_exec", 1),
-    "source": ("tcl_source", 1),
-    "load": ("tcl_load", 1),
-    "unload": ("tcl_unload", 1),
-    # Format / regex / encoding stubs.
-    "format": ("tcl_format", 4),
-    "scan": ("tcl_scan", 2),
-    "binary": ("tcl_binary", 2),
-    "regexp": ("tcl_regexp", 2),
-    "regsub": ("tcl_regsub", 2),
-    "encoding": ("tcl_encoding", 3),
-    # Event / coroutine stubs.  The arithmetic clock commands
-    # (``clock seconds`` / ``clock clicks`` / ``clock milliseconds``)
-    # are real runtime fns and route through a separate dispatcher
-    # in ``_emit_cmd_clock``; only the formatting variants surface
-    # here.
-    "after": ("tcl_after", 1),
-    "vwait": ("tcl_vwait", 1),
-    "update": ("tcl_update", 0),
-    "coroutine": ("tcl_coroutine", 2),
-    "yield": ("tcl_yield", 1),
-    "yieldto": ("tcl_yieldto", 1),
-    # Environment / metadata stubs.  ``namespace eval`` is compiled
-    # (it is a control-flow construct); the ``namespace`` entry here
-    # only catches the other subcommands (current, qualifiers, which,
-    # tail, code, delete, import, export, exists, parent, children,
-    # inscope, origin, forget, path, ensemble) that route through
-    # this dispatch.  Likewise ``interp`` covers the commands the
-    # previous ``_UNSUPPORTED_COMMANDS`` entry used to hard-trap;
-    # the stub trap is equivalent but gives a sourced site.
-    "package": ("tcl_package", 2),
-    "trace": ("tcl_trace", 2),
-    "interp": ("tcl_interp", 2),
-    "apply": ("tcl_apply", 2),
-}
-
-# String sub-command → import key
-_STRING_SUBCMD_IMPORT: dict[str, str] = {
-    "length": "tcl_string_length",
-    "index": "tcl_string_index",
-    "range": "tcl_string_range",
-    "compare": "tcl_string_compare",
-    "match": "tcl_string_match",
-    "map": "tcl_string_map",
-    "trim": "tcl_string_trim",
-    "trimleft": "tcl_string_trimleft",
-    "trimright": "tcl_string_trimright",
-    "equal": "tcl_string_equal",
-    "first": "tcl_string_first",
-    "last": "tcl_string_last",
-    "repeat": "tcl_string_repeat",
-    "reverse": "tcl_string_reverse",
-    "toupper": "tcl_string_toupper",
-    "tolower": "tcl_string_tolower",
-    "replace": "tcl_string_replace",
-}
-
-# ``string is <class> <value>`` sub-sub-command → import key
-_STRING_IS_IMPORT: dict[str, str] = {
-    "integer": "tcl_string_is_integer",
-    "alpha": "tcl_string_is_alpha",
-    "digit": "tcl_string_is_digit",
-    "space": "tcl_string_is_space",
-}
-
-# Dict sub-command → (import key, additional_arg_count after dict_var)
-_DICT_SUBCMD_IMPORT: dict[str, str] = {
-    "get": "tcl_dict_get",
-    "set": "tcl_dict_set",
-    "exists": "tcl_dict_exists",
-    "keys": "tcl_dict_keys",
-    "values": "tcl_dict_values",
-    "size": "tcl_dict_size",
-    "create": "tcl_dict_create",
-}
-
-# ``clock <subcmd>`` → import key.  Only subcommands that map to a
-# WASI-backed runtime hook are listed; ``format``/``scan`` fall through
-# to the interpreter which itself traps in the sandbox.
-_CLOCK_SUBCMD_IMPORT: dict[str, str] = {
-    "seconds": "tcl_clock_seconds",
-    "clicks": "tcl_clock_clicks",
-    "milliseconds": "tcl_clock_milliseconds",
-}
-
-# Commands that are scope declarations, compile-time-only, or CFG
-# placeholders — NOPs in WASM.
-_SCOPE_NOP_COMMANDS = frozenset(
-    {
-        "namespace",
-        "proc",
-        "package",
-        "rename",
-        # CFG placeholders — the actual logic is emitted by the loop/foreach
-        # emitters; these IRCall nodes are just iteration-setup markers.
-        "foreach",
-    }
+from ._ir import (  # noqa: E402
+    _BLOCK_I64,
+    _BLOCK_VOID,
+    _WASM_MAGIC,
+    _WASM_VERSION,
+    DiagMap,
+    DiagSite,
+    SectionId,
+    ValType,
+    WasmData,
+    WasmFunction,
+    WasmImport,
+    WasmInstruction,
+    WasmModule,
+    WasmOp,
+    _decode_leb128_signed,
+    _decode_leb128_unsigned,
+    _format_wat_instr,
+    _valtype_name,
 )
-
-# Commands that require capabilities unavailable in the WASM sandbox.
-# The codegen emits a call to the runtime ``error`` function with a
-# descriptive message so the module traps with a clear diagnostic
-# rather than silently emitting a NOP.
-# Commands that have no meaningful implementation in WASM and
-# aren't covered by the ``_CMD_RUNTIME`` stub table.  Kept as an
-# explicit trap path so (a) we reject them at compile time with a
-# clear message and (b) new users see they're intentionally
-# unimplemented rather than silently accepted.  Most commands that
-# used to live here (``exec``, ``socket``, ``interp``, ``after``,
-# etc.) have moved to the stub dispatch table — they still trap,
-# but the diag machinery attributes the trap to a source site.
-_UNSUPPORTED_COMMANDS: frozenset[str] = frozenset()
-
 
 # Parsing helpers extracted to _parsing.py — re-exported here for
 # back-compat and internal use.
